Address code review: shared ValueCategory header and fix capability names binary bloat
Two code review fixes for PR #10610:

1. Extract duplicate ValueCategory enum to shared header
   - Created slang-options-value-category.h with ValueCategory enum,
     GetValueCategory template, and SLANG_GET_VALUE_CATEGORY macros
   - Updated both slang-options.cpp and slang-options-init.cpp to
     include the shared header instead of duplicating definitions
   - Prevents silent option parsing errors from enum drift

2. Fix binary bloat from duplicate capability data tables
   - Excluded slang-capability-names.cpp from slang library build
   - File is already compiled in slang-help target
   - Prevents duplication of large static arrays (kCapabilityArray,
     kCapabilityConjunctions, kCapabilityNameInfos) from
     slang-generated-capability-defs-impl.h
